Day 6 Part 2 Complete
I do wonder if there's a better way to approach this; the simulation
took a couple seconds to run meaning that we might be getting to the
point where some more clever approaches are necessary.  That being
said, the impl worked on the first try :)

I think you could definitely implement something similar with a few
optimizations here and there that would reduce the search space.  That
being said, it isn't terrible..

Actually, running again in release, this isn't really too slow at all
with this input at around 0.25s for the run and you could totally
rayonize the simulation portion to make use of cores.

``
./target/release/d6  0.24s user 0.03s system 99% cpu 0.273 total
```