Rework existing implementations to working with wasm32-wasip1 target
This is mostly interesting for doing some side-learnings related
to wasm and doing time comparisons.  Within the sandbox, the
use of `env!("CARGO_MANIFEST_DIR")` was problematic and, really,
not necessary so just did a replace with CWD for these allowing
us to mount in `.` when as a dir when running wasmtime to get
around that issue.