Day 13 Part Two Complete
It has been a few years since having to solve a system of linear
equations.  It turns out that hasn't been a staple of my
particular professional career :shrug: -- got some hints on this
one to sort out the math as it has been a long time since
taking linear.