Day 16 Almost? Working
This implementation with memoization is working for the sample
inputs but yields a result that is too high for the full input
when submitted, so must be a problem with some case.

The original soluiton which did not have the memo/cache
took prohibitively long to run so did a first past at
optimization.  This runs fast enough, just need to get it
correct - pt.2 could maybe require something more clever
like A* but this approach seems to work (DFS).