Day 16 Part 2 Complete
Moving the path tracking into the Stete for a given path ended up
being a nice solution here rather than trying to rebuild the path
from the prev records which got messy.

The route printing is working as well; could be nice to have
a more fancy visualization or way to show the unique path
segements, but on to the next problem, shrug.